client/requirements.txt
Flask==3.0.3
Flask-WTF==1.2.2
grpcio==1.67.1
grpcio-reflection==1.67.1
grpcio-tools==1.67.1
Jinja2==3.1.4
protobuf==5.28.3
python-dotenv==1.0.1
WTForms==3.2.1
gunicorn